Refact: move files to another folder
