+++ src/components/Main/Case/Case.module.scss
