+++ src/components/Main/Discount/Discount.module.scss
