+++ src/components/Main/Specifications/Specifications.module.scss
