+++ src/components/Main/Sponsors/Sponsor.module.scss
