+++ src/components/Main/index.tsx
@@ -1,9 +1,10 @@
 import s from './Main.module.scss'
-import { Home } from '../Home'
-import { Sponsors } from '../Sponsors'
-import { Specifications } from '../Specifications'
-import { Case } from '../Case'
-import { Discount } from '../Discount'
+import { Home } from './Home'
+import { Sponsors } from './Sponsors'
+import { Specifications } from './Specifications'
+import { Case } from './Case'
+import { Discount } from './Discount'
+import { Products } from './Products'
 
 export const Main = () => {
   return (
@@ -13,6 +14,7 @@ export const Main = () => {
       <Specifications />
       <Case />
       <Discount />
+      <Products />
     </main>
   )
 }
